feat(tests): running master-diff report after each UI test
[no changelog]
