+++ ci/prepare_ui_artifacts.py
@@ -5,10 +5,11 @@ from pathlib import Path
 ROOT = Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(ROOT))
 # Needed for setup purposes, filling the FILE_HASHES dict
-from tests.ui_tests.common import TestResult, _hash_files, get_fixtures  # isort:skip
+from tests.ui_tests.common import TestResult, _hash_files  # isort:skip
+from tests.ui_tests.common import get_current_fixtures  # isort:skip
 
 
-FIXTURES = get_fixtures()
+FIXTURES = get_current_fixtures()
 
 for result in TestResult.recent_results():
     if not result.passed or result.expected_hash != result.actual_hash:
